+++ src/hal/xr809/hal_pins_xr809.c
@@ -118,7 +118,7 @@ void HAL_PIN_PWM_Stop(int pinIndex) {
 void HAL_PIN_PWM_Start(int index) {
 
 }
-void HAL_PIN_PWM_Update(int index, int value) {
+void HAL_PIN_PWM_Update(int index, float value) {
 
 	if(value<0)
 		value = 0;
